requirements.txt
py4j>=0.10.9
requests>=2.25.0
openai-whisper>=20231117
pyaudio>=0.2.11